+++ Changelog.md
@@ -31,12 +31,13 @@ HOW I DO VERSIONING:
 <!-- ### :warning: Known issues -->
 ### :ballot_box_with_check: Fixed
 - [3.1.7.1] Uncaught exception in lmcut computation
+- [3.1.7.2] Added integrity check in reading CPLEX solution
 - [3.1.7.3] Fixed issues #1,#2 and #4
 - [3.1.7.4] Swapped unordered set, which were iterated upon, for a set (determinism across platforms where possible)
+- [3.1.7.5] Fixed issue #3 and removed unreliable statistic on root node relaxation (kept relaxation and cutloop lower bounds on custom cutloop)
 <!-- ### :heavy_exclamation_mark: Changed -->
 ### :heavy_plus_sign: Added
 - Added option to add fractional callbacks to complete (TL and VE) models (aswell as custom cutloop)
-- [3.1.7.2] Added integrity check in reading CPLEX solution
 <!-- ### :x: Removed -->
 <!-- ### :curly_loop: Other -->
 <!-- ### :rocket: Performance Improvements -->